Validate and report Burmese and Illustration Glossaries
